+++ src/types.rs
@@ -3,10 +3,11 @@ use std::rc::Rc;
 
 use crate::graphics::ppu::Ppu;
 use crate::interfaces::Memory;
-use crate::processor::bus::Bus;
+use crate::processor::bus::{Bus, GraphicsBus};
 use crate::processor::memory::{Ciram, MirroredMemory, Ram, Rom};
 
 pub type SharedBus = Rc<RefCell<Bus>>;
+pub type SharedGraphicsBus = Rc<RefCell<GraphicsBus>>;
 
 pub type SharedMemory = Rc<RefCell<dyn Memory>>;
 pub type SharedRam = Rc<RefCell<Ram>>;