+++ changelog.rst
@@ -5,6 +5,8 @@ CHANGELOG
 ------
 - Performance optimizations:
   - Faster instruction set lookup with match instead of a HashMap
+  - Replace graphics bus HashMap with fields
+  - CPU RAM inside the main bus struct
 
 0.62.1
 ------